requirements.txt
asgiref==3.2.7
boto3==1.13.1
botocore==1.16.1
dj-database-url==0.5.0
django==2.2.3
django-admin-interface==0.12.2
django-ckeditor==5.9.0
django-colorfield==0.3.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
django-storages==1.9.1
django-taggit==1.3.0
django-tinymce==3.0.2
docutils==0.15.2
google-api-core==1.20.0
google-auth==1.17.0
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
gunicorn==20.0.4
jmespath==0.9.5
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
s3transfer==0.3.3
setuptools==40.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1